style(question info): moved question info into own component
